Resolve "Remove f90wrap from build process"

Closes #1397

See merge request process/process!520

+++ source/fortran/costs_2015.f90
@@ -453,17 +453,17 @@ contains
     integer :: i, j
 
     ! Enrichment variables
-    real(dp) :: product_li6
-    real(dp) :: feed_li6
-    real(dp) :: tail_li6
-    real(dp) :: feed_to_product_mass_ratio
-    real(dp) :: tail_to_product_mass_ratio
-    real(dp) :: p_v, f_v, t_v
-    real(dp) :: swu, total_swu
-    real(dp) :: mass_li
+    real(8) :: product_li6
+    real(8) :: feed_li6
+    real(8) :: tail_li6
+    real(8) :: feed_to_product_mass_ratio
+    real(8) :: tail_to_product_mass_ratio
+    real(8) :: p_v, f_v, t_v
+    real(8) :: swu, total_swu
+    real(8) :: mass_li
 
     ! First wall W coating variables
-    real(dp) :: W_density
+    real(8) :: W_density
 
     ! !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
 
@@ -614,12 +614,12 @@ contains
     integer :: i
 
     !  Divertor RH system variables
-    !real(dp) :: ITER_num_div_rh_systems
-    !real(dp) :: div_num_rh_systems_ratio
+    !real(8) :: ITER_num_div_rh_systems
+    !real(8) :: div_num_rh_systems_ratio
 
     !  First wall and blanket RH system variables
-    !real(dp) :: ITER_num_blanket_rh_systems
-    !real(dp) :: blanket_num_rh_systems_ratio
+    !real(8) :: ITER_num_blanket_rh_systems
+    !real(8) :: blanket_num_rh_systems_ratio
 
     ! !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
 
@@ -1043,8 +1043,8 @@ contains
     implicit none
 
     !  Arguments
-    real(dp), intent(in) :: x
-    real(dp), intent(out) :: v
+    real(8), intent(in) :: x
+    real(8), intent(out) :: v
 
     ! !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
 
@@ -1065,7 +1065,7 @@ contains
     integer :: n
     character(len=*), intent(in) :: descr
     character(len=5) :: vname
-    real(dp), intent(in) :: value
+    real(8), intent(in) :: value
     !  Local variables
     character(len=70) :: dum70
 
@@ -1094,7 +1094,7 @@ contains
     integer, intent(in) :: file
     character(len=*), intent(in) :: descr
     character(len=*), intent(in) :: vname
-    real(dp), intent(in) :: value
+    real(8), intent(in) :: value
     !  Local variables
     character(len=70) :: dum70
